change int16 back to int, node number can exceed 65536

+++ TomoWarp2_v2.1.320/DIC_setup.py
@@ -83,7 +83,7 @@ def DIC_setup( kinematics, data, q_data_requests, workerQueues ):
     # --- Define the extent of data we need to ask for ---
     # ----- i.e., generate extents matrix ----------------
     # 2014-10-04 EA and ET: updating extents matrix to a 4D array, with { node number }, { im_number 0,1 }, { top, bottom }, { z, y, x }
-    extents = numpy.zeros( ( kinematics.shape[0], 2, 2, 3 ), dtype=numpy.int16 )
+    extents = numpy.zeros( ( kinematics.shape[0], 2, 2, 3 ), dtype=numpy.int )
 
     #                     position            correlation_window            top extent of search window     prior displacement
     # --- Handling im1_lo
